Website/IVP/requirements.txt
absl-py==1.0.0
albumentations==1.3.0
astunparse==1.6.3
beautifulsoup4==4.11.1
cachelib==0.7.0
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cmake==3.26.3
colorama==0.4.6
crayons==0.4.0
deepface==0.0.75
efficientnet-pytorch==0.7.1
filelock==3.7.0
fire==0.4.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==1.12
gast==0.4.0
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.46.3
h5py==3.7.0
idna==3.3
imageio==2.28.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
lazy_loader==0.2
libclang==14.0.1
lit==16.0.2
Markdown==3.3.7
MarkupSafe==2.1.1
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
piexif==1.1.3
Pillow==9.5.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.8
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
stegano==0.11.1
sympy==1.11.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.0
triton==2.0.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
